refactor(providers): isolate Bedrock mantle routing
